+++ models/office.go
@@ -4,5 +4,6 @@ package models
 type Office struct {
 	ID       int    `json:"id"`
 	Name     string `json:"name"`
+	IsGlobal bool   `json:"is_global"`
 	IsActive bool   `json:"is_active"`
 }