+++ models/statistics_tic_tac_toe.go
@@ -1,5 +1,7 @@
 package models
 
+import "github.com/guregu/null"
+
 var (
 	// TicTacToeWinningCombos contains all winning combincations for tic tac toe
 	TicTacToeWinningCombos = [8][3]int{
@@ -21,15 +23,16 @@ var (
 
 // StatisticsTicTacToe struct used for storing statistics for tic tac toe
 type StatisticsTicTacToe struct {
-	ID            int `json:"id,omitempty"`
-	LegID         int `json:"leg_id,omitempty"`
-	PlayerID      int `json:"player_id,omitempty"`
-	MatchesPlayed int `json:"matches_played"`
-	MatchesWon    int `json:"matches_won"`
-	LegsPlayed    int `json:"legs_played"`
-	LegsWon       int `json:"legs_won"`
-	DartsThrown   int `json:"darts_thrown"`
-	Score         int `json:"score" `
-	NumbersClosed int `json:"numbers_closed"`
-	HighestClosed int `json:"highest_closed"`
+	ID            int      `json:"id,omitempty"`
+	LegID         int      `json:"leg_id,omitempty"`
+	PlayerID      int      `json:"player_id,omitempty"`
+	MatchesPlayed int      `json:"matches_played"`
+	MatchesWon    int      `json:"matches_won"`
+	LegsPlayed    int      `json:"legs_played"`
+	LegsWon       int      `json:"legs_won"`
+	OfficeID      null.Int `json:"office_id,omitempty"`
+	DartsThrown   int      `json:"darts_thrown"`
+	Score         int      `json:"score" `
+	NumbersClosed int      `json:"numbers_closed"`
+	HighestClosed int      `json:"highest_closed"`
 }